requirements.txt
openai
python-dotenv
numpy
chromadb>=0.4.24
tiktoken